# Improve #find_matching_token performance