--- a/Gemfile
+++ b/Gemfile
@@ -24,3 +24,4 @@ gem "activerecord-jdbcsqlite3-adapter", platform: :jruby
 gem "sqlite3", "~> 1.4", platform: %i[ruby mswin mingw x64_mingw]
 
 gem "tzinfo-data", platforms: %i[mingw mswin x64_mingw]
+gem "timecop"