Add expiration time math for `#matching_token_for`. Improves performance of the token lookup for known AR db adapters. For unknown it provides a warning and allows to add a custom SQL.

---

Some real scenario when we have ~1 600 000 expired tokens and just a few or them are still active:

```ruby
# Tests were performed for MySQL

app = Doorkeeper::Application.create!(name: "Test", redirect_uri: "https://domain.com")

Doorkeeper::AccessToken.transaction do
  500.times do 
    Doorkeeper::AccessToken.create!(application: app, expires_in: nil)
  end

  1_600_000.times do  
     Doorkeeper::AccessToken.create!(application: app, expires_in: 1)
  end
end
```

## IPS performance

```ruby
require 'benchmark/ips'

app = Doorkeeper::Application.first

Benchmark.ips do |x|
  x.report("Old lookup") do
    Doorkeeper::AccessToken.matching_token_for(app, nil, "")
  end
  
  x.report("New lookup") do
    Doorkeeper::AccessToken.matching_token_for2(app, nil, "")
  end

  x.compare!
end

# Warming up --------------------------------------
#           Old lookup     1.000  i/100ms
#           New lookup     1.000  i/100ms
# Calculating -------------------------------------
#           Old lookup      0.018  (± 0.0%) i/s -      1.000  in  55.587743s
#           New lookup      0.143  (± 0.0%) i/s -      1.000  in   7.017439s
# 
# Comparison:
#           New lookup:        0.1 i/s
#           Old lookup:        0.0 i/s - 7.92x  (± 0.00) slower
```

## Execution performance

```ruby
app = Doorkeeper::Application.first

Benchmark.bm do |x|
  x.report('Old lookup') do
    5.times { Doorkeeper::AccessToken.matching_token_for(app, nil, "") }
  end

  x.report('New lookup') do
    5.times { Doorkeeper::AccessToken.matching_token_for2(app, nil, "") }
  end
end

#              user              system      total           real
# Old lookup   258.036051   0.852471    258.888522  (289.124156)
# New lookup  0.084175      0.000020     0.084195     ( 32.517308)

```

Should fix #1485 , #1310
Relates to #1487